+++ src/renderer/api/jellyfin/jellyfin-controller.ts
@@ -258,9 +258,9 @@ export const JellyfinController: InternalControllerEndpoint = {
 
         return {
             credential: res.body.AccessToken,
-            isAdmin: Boolean(res.body.User.Policy.IsAdministrator),
+            isAdmin: Boolean(res.body.User.Policy?.IsAdministrator),
             userId: res.body.User.Id,
-            username: res.body.User.Name,
+            username: res.body.User.Name ?? body.username,
         };
     },
     createFavorite: async (args) => {